# backend/docs/swagger.yaml
basePath: /
definitions:
...
  v1.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  v1.PlaylistPageResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/v1.PlaylistResponse'
        type: array
      limit:
        type: integer
      next_cursor:
        type: string
    type: object
  v1.PlaylistPatchRequest:
    properties:
      is_public:
        type: boolean
      title:
        type: string
    type: object
  v1.PlaylistReplaceRequest:
    properties:
      is_public:
        type: boolean
      title:
        type: string
    type: object
  v1.PlaylistResponse:
    properties:
      id:
        type: string
      is_public:
        type: boolean
      owner_id:
        type: string
      title:
        type: string
    type: object
  v1.PlaylistTrackAddRequest:
    properties:
      position:
        type: integer
      track_id:
        type: string
    type: object
  v1.PlaylistTrackMoveRequest:
    properties:
      position:
        type: integer
    type: object
  v1.PlaylistTrackPageResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/v1.PlaylistTrackResponse'
        type: array
      limit:
        type: integer
      next_cursor:
        type: string
    type: object
  v1.PlaylistTrackResponse:
    properties:
      position:
        type: integer
      track:
        $ref: '#/definitions/v1.TrackResponse'
    type: object
...
  v1.TrackPageResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/v1.TrackResponse'
        type: array
      limit:
        type: integer
      next_cursor:
        type: string
    type: object
...
  v1.TrackResponse:
    properties:
      artist:
        type: string
      id:
        type: string
      size_bytes:
        type: integer
      status:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  description: Music streaming backend
  title: Chimera API
  version: "1.0"
paths:
...
  /v1/me/playlists:
    get:
      operationId: listMyPlaylists
      parameters:
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.PlaylistPageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List current user's playlists
      tags:
      - playlists
...
  /v1/playlists:
    post:
      consumes:
      - application/json
      operationId: createPlaylist
      parameters:
      - description: Playlist
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.PlaylistReplaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.PlaylistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a playlist
      tags:
      - playlists
  /v1/playlists/{playlistId}:
    delete:
      operationId: deletePlaylist
      parameters:
      - description: Playlist ID
        in: path
        name: playlistId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a playlist
      tags:
      - playlists
    get:
      operationId: getPlaylist
      parameters:
      - description: Playlist ID
        in: path
        name: playlistId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.PlaylistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get a visible playlist
      tags:
      - playlists
    patch:
      consumes:
      - application/json
      operationId: updatePlaylist
      parameters:
      - description: Playlist ID
        in: path
        name: playlistId
        required: true
        type: string
      - description: Playlist fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.PlaylistPatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.PlaylistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update playlist fields
      tags:
      - playlists
    put:
      consumes:
      - application/json
      operationId: replacePlaylist
      parameters:
      - description: Playlist ID
        in: path
        name: playlistId
        required: true
        type: string
      - description: Playlist
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.PlaylistReplaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.PlaylistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Replace a playlist
      tags:
      - playlists
  /v1/playlists/{playlistId}/tracks:
    get:
      operationId: listPlaylistTracks
      parameters:
      - description: Playlist ID
        in: path
        name: playlistId
        required: true
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.PlaylistTrackPageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: List tracks in a visible playlist
      tags:
      - playlists
    post:
      consumes:
      - application/json
      operationId: addPlaylistTrack
      parameters:
      - description: Playlist ID
        in: path
        name: playlistId
        required: true
        type: string
      - description: Track and position
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.PlaylistTrackAddRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.PlaylistTrackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a track to a playlist
      tags:
      - playlists
  /v1/playlists/{playlistId}/tracks/{trackId}:
    delete:
      operationId: removePlaylistTrack
      parameters:
      - description: Playlist ID
        in: path
        name: playlistId
        required: true
        type: string
      - description: Track ID
        in: path
        name: trackId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove a track from a playlist
      tags:
      - playlists
    patch:
      consumes:
      - application/json
      operationId: movePlaylistTrack
      parameters:
      - description: Playlist ID
        in: path
        name: playlistId
        required: true
        type: string
      - description: Track ID
        in: path
        name: trackId
        required: true
        type: string
      - description: New position
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.PlaylistTrackMoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.PlaylistTrackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Move a track within a playlist
      tags:
      - playlists
  /v1/tracks:
    get:
      description: Lists ready tracks only.
      operationId: listTracks
      parameters:
      - description: Exact artist credit (case-insensitive)
        in: query
        name: artist
        type: string
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.TrackPageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Track feed
      tags:
      - tracks
...
  /v1/users/{userId}/playlists:
    get:
      operationId: listUserPlaylists
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.PlaylistPageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: List a user's public playlists
      tags:
      - playlists
...
securityDefinitions:
  BearerAuth:
    description: JWT from /v1/auth/login. Paste only the token, or "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"